DataInputSimulator/requirements.txt
basyx-python-sdk==1.0.0
certifi==2024.7.4
charset-normalizer==3.3.2
future==1.0.0
idna==3.8
lxml==4.9.4
pip==24.2
pyecma376-2==1.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==72.1.0
six==1.16.0
urllib3==1.26.19
wheel==0.43.0
